+++ ms_lesion_segmentation/animaMSExamPreparationMSSEG2016.py
@@ -10,7 +10,7 @@ else:
 
 import os
 import shutil
-import tempfile
+import uuid
 from subprocess import call, check_output
 
 configFilePath = os.path.join(os.path.expanduser("~"), ".anima",  "config.txt")
@@ -36,8 +36,15 @@ parser.add_argument('-g', '--t1-gd', required=True, help='Path to the MS patient
 parser.add_argument('-T', '--t2', required=True, help='Path to the MS patient T2 image to register')
 parser.add_argument('-p', '--pd', required=True, help='Path to the MS patient PD image to register')
 
+parser.add_argument('-K', '--keep-intermediate-folder', action='store_true',
+                    help='Keep intermediate folder after script end')
+
 args = parser.parse_args()
-tmpFolder = tempfile.mkdtemp()
+
+tmpFolder = os.path.join(os.path.dirname(args.reference), 'ms_prepare_2016_' + str(uuid.uuid1()))
+
+if not os.path.isdir(tmpFolder):
+    os.mkdir(tmpFolder)
 
 # Anima commands
 animaPyramidalBMRegistration = os.path.join(animaDir, "animaPyramidalBMRegistration")
@@ -118,4 +125,5 @@ for i in range(0, len(listImages)):
     biasCorrectionCommand = [animaN4BiasCorrection, "-i", maskedSecondImage, "-o", outputPreprocessedFile]
     call(biasCorrectionCommand)
 
-shutil.rmtree(tmpFolder)
+if not args.keep_intermediate_folder:
+    shutil.rmtree(tmpFolder)
